requirements.txt
# Основные зависимости для Telegram бота
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0

# Веб-сервер
Flask==3.0.0
gunicorn==21.2.0

# База данных
SQLAlchemy==2.0.32
pg8000==1.31.2

# Утилиты для Render
PyYAML==6.0.1

# Дополнительные зависимости для продакшена
Werkzeug==3.0.1
